# swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","summary":"Login user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginAuthDto"},"examples":{"example_1":{"value":{"email":"admin@example.com","password":"password123"}}}}}},"responses":{"201":{"description":"Success"}},"tags":["Auth"]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","summary":"Register user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterAuthDto"},"examples":{"example_1":{"value":{"email":"user1@gmail.com","password":"12345","username":"user01","full_name":"User 1","phone":"0901212999","avatar":"","address":"","dob":"26/02/2000","role":"User"}}}}}},"responses":{"201":{"description":"Success"}},"tags":["Auth"]}},"/api/v1/auth/refresh-token":{"post":{"operationId":"AuthController_refreshTokens","summary":"Refresh token","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string"}}}}}},"responses":{"201":{"description":"Success"}},"tags":["Auth"]}},"/user":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/user/test":{"get":{"operationId":"UserController_customResponse","parameters":[],"responses":{"200":{"description":""}}}},"/user/query":{"get":{"operationId":"UserController_useQuery","parameters":[{"name":"userId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/user/param/{userId}":{"get":{"operationId":"UserController_useParam","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Movie Ticket Booking API","description":"API for booking movie tickets","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginAuthDto":{"type":"object","properties":{"email":{"type":"string","default":"admin@gmail.com"},"password":{"type":"string","default":"123456"}},"required":["email","password"]},"RegisterAuthDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"username":{"type":"string"},"full_name":{"type":"string"},"phone":{"type":"string"},"dob":{"type":"string","example":"  ","description":"Date of birth in format dd/mm/yyyy"},"address":{"type":"string"},"avatar":{"type":"string"},"role":{"type":"string","default":"User"}},"required":["email","password","username","full_name","phone","dob","address","avatar","role"]}}},"externalDocs":{"description":"/swagger/v1/swagger.json","url":"http://localhost:3000/swagger/v1/swagger.json"}}